StepCostFunction modified so that it also works for small distances between 0 and 1. Hope, that's ok...

+++ src/aima/test/search/map/MapStepCostFunctionTest.java
@@ -25,16 +25,16 @@ public class MapStepCostFunctionTest extends TestCase {
 	}
 
 	public void testCosts() {
-		assertEquals(new Double(6), mscf.calculateStepCost("A", "B", "Go"));
-		assertEquals(new Double(7), mscf.calculateStepCost("A", "C", "Go"));
-		assertEquals(new Double(5), mscf.calculateStepCost("B", "C", "Go"));
-		assertEquals(new Double(8), mscf.calculateStepCost("C", "D", "Go"));
-		assertEquals(new Double(15), mscf.calculateStepCost("B", "E", "Go"));
+		assertEquals(new Double(5), mscf.calculateStepCost("A", "B", "Go"));
+		assertEquals(new Double(6), mscf.calculateStepCost("A", "C", "Go"));
+		assertEquals(new Double(4), mscf.calculateStepCost("B", "C", "Go"));
+		assertEquals(new Double(7), mscf.calculateStepCost("C", "D", "Go"));
+		assertEquals(new Double(14), mscf.calculateStepCost("B", "E", "Go"));
 		//
-		assertEquals(new Double(6), mscf.calculateStepCost("B", "A", "Go"));
-		assertEquals(new Double(7), mscf.calculateStepCost("C", "A", "Go"));
-		assertEquals(new Double(5), mscf.calculateStepCost("C", "B", "Go"));
-		assertEquals(new Double(8), mscf.calculateStepCost("D", "C", "Go"));
+		assertEquals(new Double(5), mscf.calculateStepCost("B", "A", "Go"));
+		assertEquals(new Double(6), mscf.calculateStepCost("C", "A", "Go"));
+		assertEquals(new Double(4), mscf.calculateStepCost("C", "B", "Go"));
+		assertEquals(new Double(7), mscf.calculateStepCost("D", "C", "Go"));
 		//
 		assertEquals(new Double(1), mscf.calculateStepCost("X", "Z", "Go"));
 		assertEquals(new Double(1), mscf.calculateStepCost("A", "Z", "Go"));
